% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_object_size}
\alias{format_object_size}
\title{taken from utils:::format.object_size}
\usage{
format_object_size(x, units = "b", ...)
}
\description{
taken from utils:::format.object_size
}
\keyword{internal}

