% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gcs_auth}
\alias{gcs_auth}
\title{Authenticate with Google Cloud Storage API}
\usage{
gcs_auth(json_file)
}
\arguments{
\item{json_file}{Authentication json file you have downloaded from your Google Project}
}
\description{
Authenticate with Google Cloud Storage API
}
\details{
The best way to authenticate is to use an environment argument pointing at your authentication file.

Set the file location of your download Google Project JSON file in a \code{GCS_AUTH_FILE} argument

Then, when you load the library you should auto-authenticate

However, you can authenticate directly using this function pointing at your JSON auth file.
}
\examples{

\dontrun{
library(googleCloudStorageR)
gcs_auth("location_of_json_file.json")
}
}
