% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{gcs_first}
\alias{gcs_first}
\alias{gcs_last}
\title{Save your R session to the cloud on startup}
\usage{
gcs_first(bucket = Sys.getenv("GCS_DEFAULT_BUCKET"))

gcs_last(bucket = Sys.getenv("GCS_DEFAULT_BUCKET"))
}
\arguments{
\item{bucket}{The bucket holding your session data}
}
\description{
Place within your \code{.Rprofile} to load and save your session data automatically
}
\details{
This will save your R sessions to Google Cloud Storage.  Control which projects via where you place the \code{.Rprofile} - see \code{?Startup} for details.

If using RStudio in Google Compute Engine, this can be useful to make sure your
  RStudio project data is persistent between sessions.

You will need to authenticate first,
 the easiest way is to make sure your authentication file is
 available in environment file \code{GCS_DEFAULT_BUCKET}
}
\examples{

\dontrun{

## within your .Rprofile file
.First <- function(){
  cat("\\n# Welcome Mark! Today is ", date(), "\\n")

  googleCloudStorageR::gcs_first()
}


.Last <- function(){
  googleCloudStorageR::gcs_last()
  message("\\nGoodbye Mark at ", date(), "\\n")
}

}
}
