% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_get_bucket}
\alias{gcs_get_bucket}
\title{Get bucket info}
\usage{
gcs_get_bucket(bucket, ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL, projection = c("noAcl", "full"))
}
\arguments{
\item{bucket}{Name of a bucket}

\item{ifMetagenerationMatch}{Return only if metageneration matches}

\item{ifMetagenerationNotMatch}{Return only if metageneration does not match}

\item{projection}{Properties to return. Default noAcl omits acl properties}
}
\value{
A bucket resource object
}
\description{
Meta data about the bucket
}
\seealso{
Other bucket functions: \code{\link{gcs_create_bucket}},
  \code{\link{gcs_list_buckets}},
  \code{\link{gcs_update_bucket}}
}

