% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl.R
\name{gcs_get_bucket_acl}
\alias{gcs_get_bucket_acl}
\title{Get Bucket Access Controls}
\usage{
gcs_get_bucket_acl(bucket = gcs_get_global_bucket(), entity = "",
  entity_type = c("user", "group", "domain", "project", "allUsers",
  "allAuthenticatedUsers"))
}
\arguments{
\item{bucket}{Name of a bucket, or a bucket object returned by \link{gcs_create_bucket}}

\item{entity}{The entity holding the permission. Not needed for entity_type \code{allUsers} or \code{allAuthenticatedUsers}}

\item{entity_type}{what type of entity

Used also for when a bucket is updated}
}
\value{
Bucket access control object
}
\description{
Returns the ACL entry for the specified entity on the specified bucket
}
\seealso{
Other Access control functions: \code{\link{gcs_create_bucket_acl}},
  \code{\link{gcs_get_object_acl}},
  \code{\link{gcs_update_object_acl}}
}

