% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_get_global_bucket}
\alias{gcs_get_global_bucket}
\title{Get global bucket name}
\usage{
gcs_get_global_bucket()
}
\value{
Bucket name
}
\description{
Bucket name set this session to use by default
}
\details{
Set the bucket name via \link{gcs_global_bucket}
}
\seealso{
Other bucket functions: \code{\link{gcs_create_bucket}},
  \code{\link{gcs_create_lifecycle}},
  \code{\link{gcs_delete_bucket}},
  \code{\link{gcs_get_bucket}},
  \code{\link{gcs_global_bucket}},
  \code{\link{gcs_list_buckets}}
}
\concept{bucket functions}
