% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{gcs_get_object}
\alias{gcs_get_object}
\title{Get an object in a bucket directly}
\usage{
gcs_get_object(bucket, object_name, meta = FALSE, saveToDisk = NULL)
}
\arguments{
\item{bucket}{bucket containing the objects}

\item{object_name}{name of object in the bucket}

\item{meta}{If TRUE then get info about the object, not the object itself}

\item{saveToDisk}{Specify a filename to save directly to disk.}
}
\value{
The object or TRUE is sucessfully saved to disk.
}
\description{
This retrieves an object directly.  Differs from via a download link.
}
\details{
The object is returned in raw format, which you need to parse yourself to use in R
  using \code{httr}'s \link[httr]{content}
}
\seealso{
Other object functions: \code{\link{gcs_list_objects}}
}

