% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl.R
\name{gcs_get_object_access}
\alias{gcs_get_object_access}
\title{Check the access control settings for an object for one entity}
\usage{
gcs_get_object_access(object_name, bucket, entity = "",
  entity_type = c("user", "group", "domain", "project", "allUsers",
  "allAuthenticatedUsers"), generation = NULL)
}
\arguments{
\item{object_name}{Name of the object}

\item{bucket}{Name of a bucket}

\item{entity}{The entity holding the permission. Not needed for entity_type \code{allUsers} or \code{allAuthenticatedUsers}}

\item{entity_type}{The type of entity}

\item{generation}{If present, selects a spcfic revision of the object}
}
\description{
Returns the default object ACL entry for the specified entity on the specified bucket.
}
\seealso{
Other Access control functions: \code{\link{gcs_update_acl}}
}

