% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub.R
\name{gcs_get_service_email}
\alias{gcs_get_service_email}
\title{Get the email of service account associated with the bucket}
\usage{
gcs_get_service_email(project)
}
\arguments{
\item{project}{The project name containing the bucket}
}
\description{
Use this to get the right email so you can give it \code{pubsub.publisher} permission.
}
\details{
This service email can be different from the email in the service JSON.  Give this
\code{pubsub.publisher} permission in the Google cloud console.
}
\seealso{
Other pubsub functions: 
\code{\link{gcs_create_pubsub}()},
\code{\link{gcs_delete_pubsub}()},
\code{\link{gcs_list_pubsub}()}
}
\concept{pubsub functions}
