% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_global_bucket}
\alias{gcs_global_bucket}
\title{Set global bucket name}
\usage{
gcs_global_bucket(bucket)
}
\arguments{
\item{bucket}{bucket name you want this session to use by default, or a bucket object}
}
\value{
The bucket name (invisibly)
}
\description{
Set a bucket name used for this R session
}
\details{
This sets a bucket to a global environment value so you don't need to
supply the bucket argument to other API calls.
}
\seealso{
Other bucket functions: 
\code{\link{gcs_create_bucket}()},
\code{\link{gcs_create_lifecycle}()},
\code{\link{gcs_delete_bucket}()},
\code{\link{gcs_get_bucket}()},
\code{\link{gcs_get_global_bucket}()},
\code{\link{gcs_list_buckets}()}
}
\concept{bucket functions}
