% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{gcs_list_objects}
\alias{gcs_list_objects}
\title{List objects in a bucket}
\usage{
gcs_list_objects(bucket)
}
\arguments{
\item{bucket}{bucket containing the objects}
}
\value{
A data.frame of the objects
}
\description{
List objects in a bucket
}
\seealso{
Other object functions: \code{\link{gcs_get_object}}
}

