% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub.R
\name{gcs_list_pubsub}
\alias{gcs_list_pubsub}
\title{List pub/sub notifications for a bucket}
\usage{
gcs_list_pubsub(bucket = gcs_get_global_bucket())
}
\arguments{
\item{bucket}{The bucket for notifications}
}
\description{
List notification configurations for a bucket.
}
\details{
Cloud Pub/Sub notifications allow you to track changes to your Cloud Storage objects.
As a minimum you wil need: the Cloud Pub/Sub API activated for the project;
sufficient permissions on the bucket you wish to monitor;
sufficient permissions on the project to receive notifications;
an existing pub/sub topic; have given your service account at least \code{pubsub.publisher} permission.
}
\seealso{
\url{https://cloud.google.com/storage/docs/reporting-changes}

Other pubsub functions: \code{\link{gcs_create_pubsub}},
  \code{\link{gcs_delete_pubsub}},
  \code{\link{gcs_get_service_email}}
}
\concept{pubsub functions}
