% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsession.R
\name{gcs_load}
\alias{gcs_load}
\title{Load an R session from the Google Cloud}
\usage{
gcs_load(bucket = gcs_get_global_bucket(), file = ".RData",
  envir = .GlobalEnv)
}
\arguments{
\item{bucket}{Bucket the stored objects are in}

\item{file}{Where the files are stored}

\item{envir}{Environment to load objects into}
}
\value{
TRUE if successful
}
\description{
Load R objects that have been saved using \link{gcs_save}
}
\seealso{
Other R session data functions: \code{\link{gcs_save}},
  \code{\link{gcs_source}}
}

