% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{gcs_read}
\alias{gcs_read}
\title{Read a file from a bucket}
\usage{
gcs_read(
  file,
  bucket = gcs_get_global_bucket(),
  read_fun,
  temp_file = FALSE,
  ...
)
}
\arguments{
\item{file}{The file in the bucket}

\item{bucket}{Name of the bucket}

\item{read_fun}{Function to convert the given file to an R object.
The first argument of this function must be the file name.}

\item{temp_file}{Whether to first download the file to a temporary location,
followed by reading and then deleting the temporary file. If \code{gcs_read}
errors for unknown reason, try setting \code{temp_file = TRUE}. See details for more
information.}

\item{...}{Other arguments passed on to import_function()}
}
\value{
An \R object
}
\description{
This reads a file from a bucket into the current \R session.
}
\details{
This function is a wrapper around \code{\link{gcs_get_object}} that allows to 
read a file from a bucket into the current R session. This is done either 
directly by establishing a raw connection, or by first downloading the file 
to a temporary location at \code{\link[base]{tempdir}}, followed by reading 
it with \code{read_fun}.
}
\examples{
\dontrun{
df <- data.frame(x = 1)

gcs_write(df, file = "df.rds", write_fun = saveRDS)
gcs_read(file = "df.rds", read_fun = readRDS)

gcs_write(df, file = "df.csv", write_fun = readr::write_csv)
gcs_read(file = "df.csv", read_fun = readr::read_csv)
}

}
