% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsession.R
\name{gcs_save}
\alias{gcs_save}
\title{Save an R session to the Google Cloud}
\usage{
gcs_save(..., list = ls(all.names = TRUE, envir = parent.frame()),
  bucket = gcs_get_global_bucket(), folder = ".RData", keepCache = FALSE)
}
\arguments{
\item{...}{Names of objects to be saved (as symbols or character strings)}

\item{list}{A character vector containing the names of objects to be saved}

\item{bucket}{Bucket to store objects in}

\item{folder}{The namespace for the files to be stored in the cloud}
}
\value{
TRUE if successful
}
\description{
If objects to be saved are for use with only R consider using this
  function which saves in a compressed format.
}
\details{
\code{gcs_save(bucket = "your_bucket")} will save all objects in the workspace
  to \code{.RData} folder on Google Cloud Storage within \code{your_bucket}.

Restore the objects using \code{gcs_load(bucket = "your_bucket")}

Objects will be uploaded separately in a folder within the bucket
  using \code{.rds} format.

This will overwrite any data with the same name in folder.
}
\seealso{
Other R session data functions: \code{\link{gcs_load}}
}

