% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsession.R
\name{gcs_source}
\alias{gcs_source}
\title{Source an R script from the Google Cloud}
\usage{
gcs_source(script, bucket = gcs_get_global_bucket(), ...)
}
\arguments{
\item{script}{The name of the script on GCS}

\item{bucket}{Bucket the stored objects are in}

\item{...}{Passed to \link{source}}
}
\value{
TRUE if successful
}
\description{
Download an R script and run it immediately via \link{source}
}
\seealso{
Other R session data functions: \code{\link{gcs_load}},
  \code{\link{gcs_save_all}}, \code{\link{gcs_save_image}},
  \code{\link{gcs_save}}
}
