% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl.R
\name{gcs_update_acl}
\alias{gcs_update_acl}
\title{Update Google Cloud Storage ObjectAccessControls}
\usage{
gcs_update_acl(bucket, object, entity, entity_type = c("user", "group",
  "domian", "project", "allUsers", "allAuthenticatedUsers"),
  role = c("READER", "OWNER"))
}
\arguments{
\item{bucket}{Google Cloud Storage bucket}

\item{object}{Object to update}

\item{entity}{entity to update or add, such as an email}

\item{entity_type}{what type of entity}

\item{role}{Access permission for entity}
}
\value{
TRUE if successful
}
\description{
Requires scopes
  https://www.googleapis.com/auth/devstorage.full_control
  https://www.googleapis.com/auth/cloud-platform
}
\seealso{
\href{https://cloud.google.com/storage/docs/json_api/v1/objectAccessControls/insert}{objectAccessControls on Google API reference}
}

