% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_update_bucket}
\alias{gcs_update_bucket}
\title{Update a bucket}
\usage{
gcs_update_bucket(bucket = gcs_get_global_bucket(),
  ifMetagenerationMatch = NULL, ifMetagenerationNotMatch = NULL,
  predefinedAcl = NULL, predefinedDefaultObjectAcl = NULL,
  projection = c("noAcl", "full"), versioning = NULL, lifecycle = NULL)
}
\arguments{
\item{bucket}{Name of the bucket, or a bucket object}

\item{ifMetagenerationMatch}{Return only if metageneration matches}

\item{ifMetagenerationNotMatch}{Return only if metageneration does not match}

\item{predefinedAcl}{Apply predefined access controls to bucket}

\item{predefinedDefaultObjectAcl}{Apply predefined access controls to objects}

\item{projection}{Properties to return. Default noAcl omits acl properties}

\item{versioning}{Set if the bucket supports versioning of its objects}

\item{lifecycle}{A list of \link{gcs_create_lifecycle} objects}
}
\description{
Update a buckets metadata using PATCH semantics
}
\details{
\href{https://cloud.google.com/storage/docs/bucket-locations}{See here for details on location options}

todo: acl, cors, lifecycle, logging, versioning, website
}
\seealso{
Other bucket functions: \code{\link{gcs_create_bucket}},
  \code{\link{gcs_create_lifecycle}},
  \code{\link{gcs_delete_bucket}},
  \code{\link{gcs_get_bucket}},
  \code{\link{gcs_get_global_bucket}},
  \code{\link{gcs_global_bucket}},
  \code{\link{gcs_list_buckets}}
}

