% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{gcs_write}
\alias{gcs_write}
\title{Write an R object to a file in a bucket}
\usage{
gcs_write(x, file, bucket = gcs_get_global_bucket(), write_fun, ...)
}
\arguments{
\item{x}{The R object to write}

\item{file}{Target file in the bucket to write}

\item{bucket}{Name of the bucket}

\item{write_fun}{The function to export the object  \code{x} into a file.
Must be a function with at least two arguments where the first
argument is the R object to save and the second the name of the file
to write.}

\item{...}{Other arguments passed on to write_fun()}
}
\value{
description
}
\description{
This exports an R object into a file and uploads it to a bucket.
}
\details{
This function is a wrapper around \code{\link{gcs_upload}}. The given \R object
is first written to a temporary file located in \code{\link[base]{tempdir}}, 
followed by uploading it.
}
\examples{
\dontrun{
df <- data.frame(x = 1)
gcs_write(df, "df.csv", write_fun = write.csv)
gcs_write(df, "df.rds", write_fun = saveRDS)
}

}
