% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\name{get_storage_host}
\alias{get_storage_host}
\title{Get the Google Cloud Storage API host to use for requests}
\usage{
get_storage_host()
}
\value{
The host to use for requests (includes scheme, host and port)
}
\description{
Uses the \code{STORAGE_EMULATOR_HOST} environment variable if set, otherwise
uses the default host (the real Google Cloud Storage API).
}
