% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudStorageR.R
\docType{package}
\name{googleCloudStorageR}
\alias{googleCloudStorageR}
\alias{googleCloudStorageR-package}
\title{googleCloudStorageR}
\description{
Interact with Google Cloud Storage API in R. Part of the 'cloudyr' project.
}
