% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\name{is.storage_emulated}
\alias{is.storage_emulated}
\title{Check if the Google Cloud Storage API is emulated}
\usage{
is.storage_emulated()
}
\value{
TRUE if the Google Cloud Storage API is emulated, FALSE otherwise
}
\description{
Check if the Google Cloud Storage API is emulated
}
