% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{objects.insert}
\alias{objects.insert}
\title{Stores a new object and metadata.}
\usage{
objects.insert(Object, bucket, contentEncoding = NULL,
  ifGenerationMatch = NULL, ifGenerationNotMatch = NULL,
  ifMetagenerationMatch = NULL, ifMetagenerationNotMatch = NULL,
  name = NULL, predefinedAcl = NULL, projection = NULL)
}
\arguments{
\item{Object}{The \link{Object} object to pass to this method}

\item{bucket}{Name of the bucket in which to store the new object}

\item{contentEncoding}{If set, sets the contentEncoding property of the final object to this value}

\item{ifGenerationMatch}{Makes the operation conditional on whether the object's current generation matches the given value}

\item{ifGenerationNotMatch}{Makes the operation conditional on whether the object's current generation does not match the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the object's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the operation conditional on whether the object's current metageneration does not match the given value}

\item{name}{Name of the object}

\item{predefinedAcl}{Apply a predefined set of access controls to this object}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\seealso{
Other Object functions: \code{\link{Object}}
}

