% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_create_bucket}
\alias{gcs_create_bucket}
\title{Create a new bucket}
\usage{
gcs_create_bucket(name, projectId, location = "US",
  storageClass = c("MULTI_REGIONAL", "REGIONAL", "STANDARD", "NEARLINE",
  "COLDLINE", "DURABLE_REDUCED_AVAILABILITY"),
  predefinedAcl = c("projectPrivate", "authenticatedRead", "private",
  "publicRead", "publicReadWrite"),
  predefinedDefaultObjectAcl = c("bucketOwnerFullControl", "bucketOwnerRead",
  "authenticatedRead", "private", "projectPrivate", "publicRead"),
  projection = c("noAcl", "full"), versioning = FALSE, lifecycle = NULL)
}
\arguments{
\item{name}{Globally unique name of bucket to create}

\item{projectId}{A valid Google project id}

\item{location}{Location of bucket. See details}

\item{storageClass}{Type of bucket}

\item{predefinedAcl}{Apply predefined access controls to bucket}

\item{predefinedDefaultObjectAcl}{Apply predefined access controls to objects}

\item{projection}{Properties to return. Default noAcl omits acl properties}

\item{versioning}{Set if the bucket supports versioning of its objects}

\item{lifecycle}{A list of \link{gcs_create_lifecycle} objects}
}
\description{
Create a new bucket in your project
}
\details{
\href{https://cloud.google.com/storage/docs/bucket-locations}{See here for details on location options}
}
\seealso{
Other bucket functions: \code{\link{gcs_create_lifecycle}},
  \code{\link{gcs_delete_bucket}},
  \code{\link{gcs_get_bucket}},
  \code{\link{gcs_get_global_bucket}},
  \code{\link{gcs_global_bucket}},
  \code{\link{gcs_list_buckets}}
}
