% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buckets.R
\name{gcs_delete_bucket}
\alias{gcs_delete_bucket}
\title{Delete a bucket}
\usage{
gcs_delete_bucket(
  bucket,
  ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL
)
}
\arguments{
\item{bucket}{Name of the bucket, or a bucket object}

\item{ifMetagenerationMatch}{Delete only if metageneration matches}

\item{ifMetagenerationNotMatch}{Delete only if metageneration does not match}
}
\description{
Delete the bucket, and all its objects
}
\seealso{
Other bucket functions: 
\code{\link{gcs_create_bucket}()},
\code{\link{gcs_create_lifecycle}()},
\code{\link{gcs_get_bucket}()},
\code{\link{gcs_get_global_bucket}()},
\code{\link{gcs_global_bucket}()},
\code{\link{gcs_list_buckets}()}
}
\concept{bucket functions}
