% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gcs_download_url}
\alias{gcs_download_url}
\title{Get the download URL}
\usage{
gcs_download_url(object_name, bucket)
}
\arguments{
\item{object_name}{A vector of object names}

\item{bucket}{A vector of bucket names

bucket names should be length 1 or same length as object_name}
}
\value{
the URL for downloading objects

You may need to update access for your email to get access via \link{gcs_update_acl}
}
\description{
Create the download URL for objects in buckets
}
\seealso{
Other download functions: \code{\link{gcs_parse_download}}
}

