% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{gcs_upload}
\alias{gcs_upload}
\title{Upload a file or arbitary type}
\usage{
gcs_upload(file, bucket, type = NULL, name = "myObject")
}
\arguments{
\item{file}{filepath to what you are uploading}

\item{bucket}{bucketname you are uploading to}

\item{type}{MIME type, guessed if NULL}

\item{name}{What to call the file once uploaded. Default is "myObject"}
}
\value{
If successful, a metadata object
}
\description{
Requires scopes https://www.googleapis.com/auth/devstorage.read_write
  or https://www.googleapis.com/auth/devstorage.full_control
}

