% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{gcs_upload}
\alias{gcs_upload}
\title{Upload a file of arbitary type}
\usage{
gcs_upload(file, bucket, type = NULL, name = file)
}
\arguments{
\item{file}{data.frame, list or filepath character to what you are uploading}

\item{bucket}{bucketname you are uploading to}

\item{type}{MIME type, guessed from file extension if NULL}

\item{name}{What to call the file once uploaded. Default is the filepath.}
}
\value{
If successful, a metadata object
}
\description{
Simple Upload - for files up to 5MB to Google Cloud Storage
}
\details{
Will turn data.frames into \code{.csv} for upload via \link[utils]{write.csv}
Will turn lists into \code{.json} via \link[jsonlite]{toJSON}
}
\section{scopes}{


Requires scopes \code{https://www.googleapis.com/auth/devstorage.read_write}
  or \code{https://www.googleapis.com/auth/devstorage.full_control}
}

