% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{gcs_version_bucket}
\alias{gcs_version_bucket}
\title{Turn bucket versioning on or off, check status (default), or
list archived versions of objects in the bucket and view their generation numbers.}
\usage{
gcs_version_bucket(bucket, action = c("status", "enable", "disable",
  "list"))
}
\arguments{
\item{bucket}{gcs bucket}

\item{action}{"status", "enable", "disable", or "list"}
}
\value{
versioned_objects dataframe #only if "list" action
}
\description{
Turn bucket versioning on or off, check status (default), or
list archived versions of objects in the bucket and view their generation numbers.
}
\concept{managementAPI functions}
